수정: 파일,폴더 이동 및 정리
