+++ yu_report/yu_190805_Search.md
@@ -151,7 +151,7 @@
 
 ![검색어 입력 1차] (https://yoonucho.github.io/post_img/search1.gif)
 
-* 오브젝트를 한꺼번에 받을수는 없는건가? 
+* 오브젝트를 한꺼번에 받을수는 없는건가? - 해결
   
  * Object.keys() 메서드는 개체 고유 속성의 이름을 배열로 반환합니다. 배열 순서는 일반 반복문을 사용할 때와 같다.
 some() 메서드는 배열 안의 어떤 요소라도 주어진 판별 함수를 통과하는지 테스트한다.
@@ -161,7 +161,9 @@ some() 메서드는 배열 안의 어떤 요소라도 주어진 판별 함수를
 	- 그냥 get으로 받아옴 
 	- 검색을 하려면 페이지가 스크롤이 되어서 나와있어야 결과가 나오는거 같다. 이건 어떻게 해야하나?
 
-* 리스트와 검색바를 통합해서 만들자. 접속시에는 /page였다가 검색바를 클릭했을때 post로 /search로 변환
-   - 통합을 하면 페이지가 바뀔 필요가 없을거 같다. 우선 기존 리스트와 통합을 했는데 스토어로(MobX) 해둔거라 연결을 하였으나
-	 서치바 부분은 잘안되어서 스토어없이 유지해서 통합은 했음
-	
+* 리스트와 검색바를 통합해서 만들자. 접속시에는 /page였다가 검색바를 클릭했을때 post로 /search로 변환?
+   - 통합을 하면 페이지가 바뀔 필요가 없을거 같다. 우선 기존 리스트와 통합을 하고 검색바 로직을 분류 (MobX)
+
+* 달력을 입력했을때 데이터가 나오게끔 설정하자.
+   -  프론트에서 캘린더를 클릭했을때 데이터가 백엔드로 받을수있게 해야하는건가?
+   -  주소에 파라미터가 있는데 그주소를 post로 받아올 수 있는가에 대한 의문
